+++ demo/blocks_inputs/run.py
@@ -10,7 +10,7 @@ Hello friends
 
 with gr.Blocks() as demo:
     txt = gr.Textbox(label="Input", lines=5)
-    txt_2 = gr.Textbox(label="Output")
+    txt_2 = gr.Textbox(label="Output-Interactive")
     txt_3 = gr.Textbox(str, label="Output", interactive=False)
     btn = gr.Button("Submit")
     btn.click(lambda a: a, inputs=[txt], outputs=[txt_2])
