+++ demo/blocks_simple_squares/config.json
@@ -1,37 +1,74 @@
 {
+  "version": "3.0.6",
   "mode": "blocks",
+  "dev_mode": true,
   "components": [
     {
       "id": 1,
       "type": "variable",
       "props": {
-        "default_value": 0,
+        "value": 0,
         "show_label": true,
         "name": "variable",
-        "css": {}
+        "visible": true,
+        "style": {}
       }
     },
     {
       "id": 2,
       "type": "number",
       "props": {
-        "default_value": 0.0,
+        "value": 0.0,
         "show_label": true,
         "name": "number",
-        "css": {}
+        "visible": true,
+        "style": {
+          "text_color": "blue",
+          "container_bg_color": "yellow"
+        }
       }
     },
     {
       "id": 3,
       "type": "button",
       "props": {
-        "default_value": "Next Square",
+        "value": "Next Square",
+        "variant": "secondary",
         "name": "button",
-        "css": {}
+        "visible": true,
+        "elem_id": "btn",
+        "style": {
+          "rounded": false,
+          "bg_color": "purple"
+        }
+      }
+    },
+    {
+      "id": 4,
+      "type": "variable",
+      "props": {
+        "value": {
+          "a": "a"
+        },
+        "show_label": true,
+        "name": "variable",
+        "visible": true,
+        "style": {}
+      }
+    },
+    {
+      "id": 5,
+      "type": "json",
+      "props": {
+        "show_label": true,
+        "name": "json",
+        "visible": true,
+        "style": {}
       }
     }
   ],
   "theme": "default",
+  "css": "#btn {color: red}",
   "enable_queue": false,
   "layout": {
     "id": 0,
@@ -44,6 +81,12 @@
       },
       {
         "id": 3
+      },
+      {
+        "id": 4
+      },
+      {
+        "id": 5
       }
     ]
   },
@@ -54,12 +97,17 @@
       ],
       "trigger": "click",
       "inputs": [
-        1
+        1,
+        4
       ],
       "outputs": [
         1,
-        2
+        2,
+        4,
+        5
       ],
+      "backend_fn": true,
+      "js": null,
       "status_tracker": null,
       "queue": null
     }
