+++ demo/blocks_xray/config.json
@@ -1,5 +1,7 @@
 {
+  "version": "3.0.6",
   "mode": "blocks",
+  "dev_mode": true,
   "components": [
     {
       "id": 1,
@@ -7,28 +9,33 @@
       "props": {
         "value": "<h1>Detect Disease From Scan</h1>\n<p>With this model you can lorem ipsum</p>\n<ul>\n<li>ipsum 1</li>\n<li>ipsum 2</li>\n</ul>\n",
         "name": "markdown",
-        "css": {}
+        "visible": true,
+        "style": {}
       }
     },
     {
       "id": 2,
       "type": "checkboxgroup",
       "props": {
-        "choices": ["Covid", "Malaria", "Lung Cancer"],
+        "choices": [
+          "Covid",
+          "Malaria",
+          "Lung Cancer"
+        ],
         "value": [],
-        "name": "checkboxgroup",
         "label": "Disease to Scan For",
         "show_label": true,
         "name": "checkboxgroup",
-        "css": {}
+        "visible": true,
+        "style": {}
       }
     },
     {
       "id": 3,
       "type": "tabs",
       "props": {
-        "css": {},
-        "value": true
+        "visible": true,
+        "style": {}
       }
     },
     {
@@ -36,8 +43,8 @@
       "type": "tabitem",
       "props": {
         "label": "X-ray",
-        "css": {},
-        "value": true
+        "visible": true,
+        "style": {}
       }
     },
     {
@@ -45,8 +52,8 @@
       "type": "row",
       "props": {
         "type": "row",
-        "css": {},
-        "value": true
+        "visible": true,
+        "style": {}
       }
     },
     {
@@ -56,18 +63,21 @@
         "image_mode": "RGB",
         "source": "upload",
         "tool": "editor",
-        "value": null,
+        "streaming": false,
+        "show_label": true,
         "name": "image",
-        "css": {}
+        "visible": true,
+        "style": {}
       }
     },
     {
       "id": 7,
       "type": "json",
       "props": {
-        "value": "\"\"",
+        "show_label": true,
         "name": "json",
-        "css": {}
+        "visible": true,
+        "style": {}
       }
     },
     {
@@ -75,8 +85,10 @@
       "type": "button",
       "props": {
         "value": "Run",
+        "variant": "secondary",
         "name": "button",
-        "css": {}
+        "visible": true,
+        "style": {}
       }
     },
     {
@@ -85,7 +97,8 @@
       "props": {
         "cover_container": true,
         "name": "statustracker",
-        "css": {}
+        "visible": true,
+        "style": {}
       }
     },
     {
@@ -93,8 +106,8 @@
       "type": "tabitem",
       "props": {
         "label": "CT Scan",
-        "css": {},
-        "value": true
+        "visible": true,
+        "style": {}
       }
     },
     {
@@ -102,8 +115,8 @@
       "type": "row",
       "props": {
         "type": "row",
-        "css": {},
-        "value": true
+        "visible": true,
+        "style": {}
       }
     },
     {
@@ -113,18 +126,21 @@
         "image_mode": "RGB",
         "source": "upload",
         "tool": "editor",
-        "value": null,
+        "streaming": false,
+        "show_label": true,
         "name": "image",
-        "css": {}
+        "visible": true,
+        "style": {}
       }
     },
     {
       "id": 13,
       "type": "json",
       "props": {
-        "value": "\"\"",
+        "show_label": true,
         "name": "json",
-        "css": {}
+        "visible": true,
+        "style": {}
       }
     },
     {
@@ -132,25 +148,46 @@
       "type": "button",
       "props": {
         "value": "Run",
+        "variant": "secondary",
         "name": "button",
-        "css": {}
+        "visible": true,
+        "style": {}
       }
     },
     {
       "id": 15,
       "type": "statustracker",
       "props": {
-        "lines": 1,
-        "placeholder": null,
-        "value": "",
-        "name": "textbox",
-        "label": null,
-        "css": {},
-        "interactive": null
+        "cover_container": true,
+        "name": "statustracker",
+        "visible": true,
+        "style": {}
+      }
+    },
+    {
+      "id": 16,
+      "type": "button",
+      "props": {
+        "value": "Upload Results",
+        "variant": "secondary",
+        "name": "button",
+        "visible": true,
+        "style": {}
+      }
+    },
+    {
+      "id": 17,
+      "type": "statustracker",
+      "props": {
+        "cover_container": false,
+        "name": "statustracker",
+        "visible": true,
+        "style": {}
       }
     }
   ],
   "theme": "default",
+  "css": null,
   "enable_queue": false,
   "layout": {
     "id": 0,
