+++ ui/packages/app/test/input-output.spec.ts
@@ -13,7 +13,6 @@ function mock_demo(page: Page, demo: string) {
 
 function mock_api(page: Page, body: Array<unknown>) {
 	return page.route("**/api/predict/", (route) => {
-		console.log("***********", route)
 		const id = JSON.parse(route.request().postData()!).fn_index;
 		return route.fulfill({
 			headers: {
@@ -28,17 +27,15 @@ function mock_api(page: Page, body: Array<unknown>) {
 
 test("a component acts as both input and output", async ({ page }) => {
 	await mock_demo(page, "input-output");
-	await mock_api(page, [["world hello"]]);
+	await mock_api(page, [["tset"]]);
 	await page.goto("http://localhost:3000");
 
 	const textbox = await page.locator("label:has-text('Input-Output')");
 	const button = await page.locator("button");
 
-	await textbox.fill("hello world");
-
+	await textbox.fill("test");
 	await Promise.all([button.click(), page.waitForResponse("**/api/predict/")]);
-
 	await expect(await page.inputValue("label:has-text('Input-Output')")).toEqual(
-		"world hello"
+		"tset"
 	);
 });
