--- ui/packages/app/test/kitchen-sink.spec.ts
@@ -1,32 +1,0 @@
-import { test, expect, Page } from "@playwright/test";
-
-function mock_demo(page: Page, demo: string) {
-	return page.route("**/config", (route) => {
-		return route.fulfill({
-			headers: {
-				"Access-Control-Allow-Origin": "*"
-			},
-			path: `../../../demo/kitchen_sink/config.json`
-		});
-	});
-}
-
-function mock_api(page: Page, body: Array<unknown>) {
-	return page.route("**/api/predict/", (route) => {
-		const id = JSON.parse(route.request().postData()!).fn_index;
-		return route.fulfill({
-			headers: {
-				"Access-Control-Allow-Origin": "*"
-			},
-			body: JSON.stringify({
-				data: body[id]
-			})
-		});
-	});
-}
-
-test("renders the correct elements", async ({ page }) => {
-	await mock_demo(page, "kitchen_sink");
-	await page.goto("http://localhost:3000");
-});
-
